requirements.txt
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.7
Telethon==1.19.1
